requirements.txt
# Core Django Framework - Latest Stable Version
Django==5.1.14

# Django Extensions
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djoser==2.3.0
social-auth-app-django==5.6.0

# CKEditor
django-ckeditor==6.7.3

# Timezone handling
pytz==2024.2

# Image Processing
Pillow==11.1.0
mutagen==1.47.0

# Database
psycopg2-binary==2.9.10

# Password Hashing
argon2-cffi==23.1.0

# Payment Processing
stripe==11.5.0
braintree==4.31.0

# JWT Authentication
PyJWT==2.10.1

# AWS/Storage
boto3==1.35.95
django-storages==1.14.4

# Utilities
markdown==3.7
gunicorn==23.0.0
whitenoise==6.8.2

# HTTP Requests
requests==2.32.5